trace_dumper: Fix get_file_size() prototype

+++ trace_dumper/filesystem.h
@@ -1,7 +1,7 @@
 #ifndef __FILESYSTEM_H__
 #define __FILESYSTEM_H__
 
-int get_file_size(const char *filename);
+long long get_file_size(const char *filename);
 long long free_bytes_in_fs(const char *mnt);
 
 #endif
